+++ client/my-sites/site-settings/traffic/index.js
@@ -8,13 +8,20 @@ import page from 'page';
  */
 import controller from './controller';
 import mySitesController from 'my-sites/controller';
+import settingsController from 'my-sites/site-settings/settings-controller';
 
 const redirectToTrafficSection = ( context ) => {
 	page.redirect( '/settings/traffic/' + ( context.params.site_id || '' ) );
 };
 
 export default function() {
-	page( '/settings/traffic/:site_id', mySitesController.siteSelection, mySitesController.navigation, controller.traffic );
+	page(
+		'/settings/traffic/:site_id',
+		mySitesController.siteSelection,
+		mySitesController.navigation,
+		settingsController.siteSettings,
+		controller.traffic
+	);
 
 	// redirect legacy urls
 	page( '/settings/analytics/:site_id?', redirectToTrafficSection );
